Complete Exercise 3.3
Create a new kind for plants, and add a few desert-appropriate examples
to some exterior rooms. The player shouldn't be able to take plants, and
they should have a new property that describes whether they are dead or
alive. Retrofit any appropriate exiting objects (like the tumbleweed) to
fall into this class.