Complete Exercise 4.1
Try creating a new room north of Base of the Tower called Against the
Fence. Give it a description and populate it with a few items. Keep in
mind the room description's dual purpose: to tell the story, and to
assist with the game. Decide whether you want to mention any fixed in
place items in the room within your description, and adjust their
properties accordingly so the player will not see them described twice.
Remember BENT (Bracket Every Notable Thing) and add any newly created
things to the story world.