Describe scenery in Base of the Tower
Complete the second round of story building prompts in Chapter 4
(Describing the Story World) between Exercise 4.1 and Exercise 4.2.

The topics covered in this section of the chapter include:

* Descriptions for Things

NOTE: When a kind is marked scenery and you add a description for the
item in source text, the description only appears if the player examines
the item.