Complete Exercise 4.2
Try writing descriptions for some of the scenery items you created in
the last chapter's exercises. Keep in mind that your description should
not introduce extraneous nouns, and should clearly indicate to the
player whether or not this item merits further investigation. Remember
too that object descriptions are a great way to set tone and reveal
small pieces of story.