Complete Exercise 4.3
Try adding an initial appearance to a portable object somewhere in the
story world. Recompile and see how the object is described when you
first encounter it, after picking it up and dropping it, and after
leaving and returning to its location. You should be able to learn
something useful about the behavior of initial appearance.

NOTE: When an item is bracketed using BENT, the initial appearance does
not display. The initial appearance description appears the first time
you encounter the object. After you move it to another location, the
description reverts to Inform's default manner of describing things that
are not bracketed in the room (e.g., You can see a ___ here).