Complete Exercise 4.4
Can you use lists and Plurality to modify the initial appearance of the
pickup truck to list its contents in a natural sounding way?

NOTE: The Plurality extension by Emily Short is now deprecated. The
functionality provided by the extension is now build into Inform 7 and
described in Chapter 14 of the documentation (Adaptive Text and
Responses).