Create definitions to support lighting conditions
A definition creates a new adjective that will be either true or false
on any given turn according to the exact parameters we specify.

Define three types of lighting conditions that apply to rooms: lit
brightly, lit by headlights, and lit by flashlight.

The topics covered in this section of the chapter include:

* Conditional Descriptions
* Conditions
* Defining New Adjectives
* Using Definitions