Document notes and references from Chapter 05
Chapter 05 introduces conventions for creating custom actions and
rulebooks and customizing and unlisting existing rules.

This chapter covers a lot of syntax, most in brief mentions and short
exercises.

The list of common verbs provided in the chapter is extensive but it is
not comprehensive. For the comprehensive list, see the alphabetical
Actions Index.