Complete Exercise 6.3
Pick a few of the extensions above and add them to Sand-dancer.
Recompile and experiment with how they change the experience of play.
You can read the documentation for an installed extension in the
Documentation panel to find out more specifically what it does and
whether Sand-dancer might benefit from it.