Chapter 07: Breaking and Entering
Begin to modify the player's entry into the building by so the player is
required to find a tool to break in.

Remove the door that allowed entry into the building.

Add the ability to break a window with an object.

NOTE: This implementation leaves a bug in place. The player can break
the window but cannot enter the building.