Chapter 07: Exercise 7.2
The examples above create temporary list variables; based on your
knowledge of regular variables, how would you create a list variable
attached to a thing? A global list variable?

NOTE: I burned a lot of time on this exercise before checking the
solution at the end of the chapter. It is commented in source because it
does not compile.