Chapter 07: Numbers
Several headings in this sections required no changes to source text,
including:

* Using Lists
* Repeating through Lists
* Numbers and Randomness
* Kinds of Numbers

In the Numbers section, we revisit the pack of cigarettes and create a
variable to track the number of cigarettes our protagonist has smoked.

While testing this, I realized a player might try a command that results
in a bug and worked through allowing the player to attempt to remove the
boards from the boarded-up door. Tests were added to confirm the source
text works as expected.