Chapter 07: Reaching the Desert
Here I begin to deviate from the text in the implementation of the story
world.

In Chapter 7, we are instructed to expand The Open Desert so it becomes
and endless expanse.

At this point when I built my initial version of this game, I started to
explore the implementation of the compass in Emily Short's Bronze. I
found that the use of color to indicate visited and unexplored
directions caused a memory leak that wouldn't allow the game to compile,
but I worked around that using bold and normal type to indicate those
states.