Chapter 10: Exercise 10.2
Knock's cigarettes are currently implemented with a number variable
count attached to the pack of cigarettes thing. How would you instead
create each cigarette as a specific object in the story world? What
would be some of the advantages and disadvantages of doing so?

NOTE: I was immediately curious about how to implement cigarettes as a
kind inside a container called pack of smokes. I completed this exercise
around the same time cigarettes were introduced in the game.

This commit:

* Adds a testing command for cigarettes
* Adds a heading to separate the implementation of cigarettes from the
  implementation of packs of cigarettes