Chapter 10: Exercise 10.3
How would you write an activity rule that would visually flag the
objects you haven't written a description for? Or visually mark anything
portable, to be sure you don't have scenery items the player could walk
off with? Can you use the later trick to find something in the building
that's portable but shouldn't be?